+++ security/src/main/java/org/xipki/security/OIDs.java
@@ -1013,6 +1013,21 @@ public class OIDs {
     public static final ASN1ObjectIdentifier id_microsoft_objectSid =
         initOid("1.3.6.1.4.1.311.25.2.1", "Microsoft-ObjectSID");
 
+    // OIDs defined in Chinese GM/T 0015-2023
+    private static final String id_gmt0012_ext = "1.2.156.10260.4";
+
+    public static final ASN1ObjectIdentifier id_cn_residentIdCardNumber =
+        initOid(id_gmt0012_ext + ".1.1.1", "CN-ResidentIdCardNumber");
+
+    public static final ASN1ObjectIdentifier id_cn_passportNumber =
+        initOid(id_gmt0012_ext + ".1.1.2", "CN-PassportNumber");
+
+    public static final ASN1ObjectIdentifier id_cn_socialInsuranceNumber =
+        initOid(id_gmt0012_ext + ".1.2", "CN-SocialInsuranceNumber");
+
+    public static final ASN1ObjectIdentifier id_cn_UnifiedSocialCreditCode =
+        initOid(id_gmt0012_ext + ".4.4", "CN-UnifiedSocialCreditCode");
+
   }
 
   /**